change image deep learning tool name and fix tool tests
