+++ galaxy_ml/tools/search_model_validation.xml
@@ -67,7 +67,6 @@
             #if $search_schemes.options.cv_selector.selected_cv in ['GroupKFold', 'GroupShuffleSplit', 'LeaveOneGroupOut', 'LeavePGroupsOut']
             --groups '$search_schemes.options.cv_selector.groups_selector.infile_g'
             #end if
-            #if $save != 'save_no_fit'
 
         ]]>
     </command>
