+++ galaxy_ml/tools/test-data/image_dl_predictions01.tabular
@@ -1,0 +1,101 @@
+0	1	2	3	4	5	6	7	8	9
+0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0
+0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0
+0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0
+0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0
+0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0
+0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0
+0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0
+1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0
+1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0
+0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0
+1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0
+1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0
+0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0
+0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0
+1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0
+1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0
+0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0
+0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0
+0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0
+1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0
+1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0
+0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0
+0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0
+0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0
+0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0
+0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0
+1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0
+0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0
+0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	1.0	0.0	0.0	0.0
+0.0	0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	1.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0
+0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	0.0	1.0
